+++ core/src/main/java/net/cofcool/chaos/server/core/aop/AbstractApiInterceptor.java
@@ -6,20 +6,23 @@ import java.util.Collection;
 import net.cofcool.chaos.server.common.core.Page;
 import net.cofcool.chaos.server.common.core.ServiceException;
 import net.cofcool.chaos.server.common.security.Auth;
-import net.cofcool.chaos.server.common.security.ExcludeType;
 import net.cofcool.chaos.server.common.security.User;
 import net.cofcool.chaos.server.common.security.UserRole;
 import net.cofcool.chaos.server.common.util.BeanUtils;
 import net.cofcool.chaos.server.core.annotation.Api;
 import net.cofcool.chaos.server.core.annotation.ApiVersion;
 import net.cofcool.chaos.server.core.annotation.DataAuthExclude;
+import net.cofcool.chaos.server.core.annotation.DataAuthExclude.ExcludeMode;
 import net.cofcool.chaos.server.core.annotation.Scanned;
 import net.cofcool.chaos.server.core.annotation.scanner.BeanResourceHolder;
+import net.cofcool.chaos.server.core.config.ChaosProperties;
 import net.cofcool.chaos.server.core.support.ExceptionCodeInfo;
 import org.aopalliance.intercept.MethodInvocation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.util.StringUtils;
 
 /**
  * 接口拦截, 包括 {@link User} 数据注入, 设备检查, 版本对比以及覆盖ID等, 通过拦截<b>Controller</b>实现,  被代理类需使用 {@link Scanned} 注解。
@@ -36,6 +39,39 @@ public abstract class AbstractApiInterceptor extends AbstractScannedMethodInterc
 
     protected final Logger log = LoggerFactory.getLogger(getClass());
 
+    /**
+     * 注入到参数的数据的key
+     */
+    private String[] definedCheckedKeys;
+
+    /**
+     * 项目版本号
+     */
+    private int version;
+
+    public int getVersion() {
+        return version;
+    }
+
+    public void setVersion(int version) {
+        this.version = version;
+    }
+
+    public String[] getDefinedCheckedKeys() {
+        return definedCheckedKeys;
+    }
+
+    public void setDefinedCheckedKeys(String[] definedCheckedKeys) {
+        this.definedCheckedKeys = definedCheckedKeys;
+    }
+
+    @Autowired
+    public void setupProperties(ChaosProperties properties) {
+        definedCheckedKeys = StringUtils
+            .delimitedListToStringArray(properties.getAuth().getCheckedKeys(), ",");
+        version = properties.getDevelopment().getVersion();
+    }
+
     private void checkApi(MethodInvocation invocation) {
         checkApi(invocation, getUser());
     }
@@ -46,16 +82,6 @@ public abstract class AbstractApiInterceptor extends AbstractScannedMethodInterc
      */
     protected abstract User getUser();
 
-    /**
-     * 项目当前版本是否大于apiVersion定义的版本
-     * @param apiVersion apiVersion
-     */
-    protected abstract boolean compareApiVersion(ApiVersion apiVersion);
-
-    /**
-     * 注入到参数的数据的key
-     */
-    protected abstract String[] getDefinedAuthDataKeys();
 
     protected void checkApi(MethodInvocation invocation, User user) {
         Api api = BeanResourceHolder.getAnnotation(invocation.getMethod(), Api.class, true);
@@ -105,10 +131,17 @@ public abstract class AbstractApiInterceptor extends AbstractScannedMethodInterc
         }
     }
 
+    /**
+     * 项目当前版本是否大于apiVersion定义的版本
+     * @param apiVersion apiVersion
+     */
+    protected boolean compareApiVersion(ApiVersion apiVersion) {
+        return version > apiVersion.value();
+    }
+
     protected void setupAuth(User user, Object[] args, DataAuthExclude authExclude) {
         if (user != null && user.getDetail() != null) {
             Auth data = user.getDetail();
-            String[] methods = getDefinedAuthDataKeys();
             for (Object arg : args) {
                 // inject user
                 if (arg instanceof User) {
@@ -118,8 +151,8 @@ public abstract class AbstractApiInterceptor extends AbstractScannedMethodInterc
 
                 // inject user properties
                 if (arg != null) {
-                    for (String property : methods) {
-                        if (checkExclude(authExclude, ExcludeType.ALL) || checkExclude(authExclude, property)) {
+                    for (String property : getDefinedCheckedKeys()) {
+                        if (checkExclude(authExclude, property)) {
                             continue;
                         }
 
@@ -134,7 +167,7 @@ public abstract class AbstractApiInterceptor extends AbstractScannedMethodInterc
         }
     }
 
-    private void setTheId(Object obj, Auth authData, String property) throws IllegalAccessException, NullPointerException, InvocationTargetException {
+    protected void setTheId(Object obj, Auth authData, String property) throws IllegalAccessException, NullPointerException, InvocationTargetException {
         Object value = BeanUtils.getPropertyDescriptor(authData.getClass(), property).getReadMethod().invoke(authData);
         if (BeanUtils.checkNullOrZero(value).isPresent()) {
             if (obj instanceof Page) {
@@ -146,16 +179,11 @@ public abstract class AbstractApiInterceptor extends AbstractScannedMethodInterc
     }
 
     private boolean checkExclude(DataAuthExclude originExclude, String excludeProperty) {
-        if (originExclude != null) {
-            try {
-                ExcludeType authTypeInstance = originExclude.type().getConstructor().newInstance();
-                for (String p : originExclude.value()) {
-                    if (authTypeInstance.getValue(p).compareTo(authTypeInstance.getValue(excludeProperty)) == 0) {
-                        return true;
-                    }
+        if (originExclude != null && originExclude.mode().equals(ExcludeMode.always)) {
+            for (String val : originExclude.value()) {
+                if (val.equals(excludeProperty)) {
+                    return true;
                 }
-            } catch (InstantiationException | IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
-                throw new IllegalStateException("can not create ExcludeType instance", e);
             }
         }
 
@@ -171,4 +199,5 @@ public abstract class AbstractApiInterceptor extends AbstractScannedMethodInterc
     protected void doBefore(MethodInvocation invocation) {
         checkApi(invocation);
     }
+
 }